+++ packages/riverpod/lib/riverpod.dart
@@ -1,5 +1,12 @@
 export 'src/common.dart'
-    show Create, AsyncValue, AsyncData, $AsyncValueCopyWith, $AsyncValue;
+    show
+        Create,
+        AsyncValue,
+        AsyncData,
+        AsyncLoading,
+        AsyncError,
+        $AsyncValueCopyWith,
+        $AsyncValue;
 export 'src/framework/framework.dart'
     show
         AlwaysAliveProviderBase,